Card Details makeover
Updated the card details to include badges for the cabin type as well as other adjustments to the pagination and search functions. Added generic search by name filter. Still need to adjust some colors for Fluorite on the Card details page but generally happy with the look and feel + functionality.