Filter fix + back to top button
Fixed issue where the parameter being passed in the url to the card detail page was called "cabin" and we were losing any selected cabin filter when returning to grid as a result. Additionally added a back to top button that appears once a user begins scrolling on both desktop and mobile.